requirements.txt
pyyaml==5.4.1
pytest==6.2.4
pytest-mock==3.6.0
web3[tester]==5.18.0
pytest-mock==3.6.0
python-dotenv==0.17.1
dump-env==1.3.0
